Started abstracting content from each column to independent components - Created new sub folder `panels` to contain the content for each "panel" of the modal to better isolate the code. - Created new `ActivityPanel` and `StudentPanel` components - Migrated the student panel code to the `StudentPanel` - Implemented new `StudentPanel` in `StudentModal` and cleaned up now unused imports.
