+++ client/src/pages/Dashboard/components/StudentModal/panels/ActivityPanel.js
@@ -1,0 +1,3 @@
+export const ActivityPanel = () => {
+    
+}
