+++ client/src/pages/Dashboard/components/StudentModal/panels/index.js
@@ -1,0 +1,2 @@
+export { ActivityPanel } from "./ActivityPanel";
+export { StudentPanel } from "./StudentPanel";
